Move bean out of test class for invalid app test
When a deployment uses ShouldThrowException, arquillian does not include
the arquillian support classes in the deployed app, including the
Arquillian class which the test extends.

In this case, if arquillian is not on the class path on the server, this
test will fail because the test bean cannot be loaded because it depends
on the Arquillian class and so doesn't cause the expected deployment
exception.
